+++ src/idtrackerai/start/arg_parser.py
@@ -247,7 +247,11 @@ def get_parser(defaults: dict | None = None) -> ArgumentParser:
         " used to train the CNN in each accumulation step",
         type=int,
     )
-
+    add_argument(
+        "bounding_box_images_on_ram",
+        "If true, bounding box images, a middle step to generate the identification images, will be kept in RAM until no longer needed. Else, they are saved in disk and loaded when needed",
+        type=Bool,
+    )
     return parser